1. Updated KataItemsDatabaseAccessMockData.GetSpecialOfferPrice to return a decimal instead of an int.
2. Added 2 additional tests to the UnitTests project.

+++ SortedCheckoutKata/Database/KataItemsDatabaseAccess/IKataItemsDatabaseAccess.cs
@@ -4,7 +4,7 @@
     {
         decimal CalculateTotalPrice(string sku, int itemQty);
         int GetItemSpecialOfferQty(string sku);
-        int GetSpecialOfferPrice(string sku);
+        decimal GetSpecialOfferPrice(string sku);
         decimal GetUnitPrice(string sku);
     }
 }
